% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louland_param.R
\docType{data}
\name{louland_param}
\alias{louland_param}
\title{Parameters for the fictional island Louland}
\format{
A table with 7 rows and 4 columns.
\describe{
\item{lc_id}{Land cover ID.}
\item{lc}{Land cover code.}
\item{w}{Land cover area weight.}
\item{hex}{Polygon geometry.}
}
}
\source{
Created with \code{NLMR}, \code{terra}, \code{stars} and \code{smoothr}.
}
\usage{
louland_param
}
\description{
Louland is a fictional island created in the middle of the Atlantic Ocean
and designed for practicing National Forest Inventory sampling and data
analysis. This table contains Louland's land cover codes, their area
weight and color code for mapping. The land cover codes represent:
EV:Evergreen Forest, MD:Mixed-Deciduous Forest, DD:Deciduous Forest,
WL: Other Woodlands, MG:Mangrove, NF:Non-Forest, and WA:Water.
For more information see:
https://www.fao.org/national-forest-monitoring/areas-of-work/nfi/modules/en/
}
\examples{
louland_param
}
\keyword{datasets}
