% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_species.R
\docType{data}
\name{raw_species}
\alias{raw_species}
\title{List of 1722 mock tree species names.}
\format{
A data frame with 1722 rows and and 3 variables:
\describe{
\item{sp_id}{Unique identifier of tree species.}
\item{sp_name}{Tree species mock genus and species names.}
\item{genus}{Tree mock genus names.}
}
}
\source{
Anonymized tropical forest inventories.
}
\usage{
raw_species
}
\description{
A dataset containing mock tree species name mimicking real species names.
These species were identified in the forest inventories and their IDs
recorded in the raw_tree table. The data should be used for educational
purposes only.
}
\examples{
  raw_species
}
\keyword{datasets}
